+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Now playlists with more than 100 videos return all the videos. Thanks to @ATiltedTree.
 - Implemented `ChannelAboutPage`, check the tests their usage.
 - Implement filters for `search.getVideos`. See `filter` getter.
+- Now video's from search queries return the channel id.
+- Implemented publishDate for videos. Thanks to @mymikemiller , PR: #115.
 
 ## 1.8.0
 - Fixed playlist client.